+++ packages/vue/src/components/IonTabBar.ts
@@ -7,6 +7,7 @@ import { h, defineComponent, getCurrentInstance, inject } from "vue";
 interface TabState {
   activeTab?: string;
   tabs: { [k: string]: Tab };
+  hasRouterOutlet?: boolean;
 }
 
 interface Tab {
@@ -37,6 +38,7 @@ export const IonTabBar = defineComponent({
     /* eslint-disable @typescript-eslint/no-empty-function */
     _tabsWillChange: { type: Function, default: () => {} },
     _tabsDidChange: { type: Function, default: () => {} },
+    _hasRouterOutlet: { type: Boolean, default: false },
     /* eslint-enable @typescript-eslint/no-empty-function */
   },
   data() {
@@ -49,10 +51,11 @@ export const IonTabBar = defineComponent({
     };
   },
   updated() {
-    this.setupTabState(inject("navManager"));
+    this.setupTabState(inject("navManager", null));
   },
   methods: {
     setupTabState(ionRouter: any) {
+      const hasRouterOutlet = this.$props._hasRouterOutlet;
       /**
        * For each tab, we need to keep track of its
        * base href as well as any child page that
@@ -74,24 +77,73 @@ export const IonTabBar = defineComponent({
 
         /**
          * Passing this prop to each tab button
+         * lets it be aware of the presence of
+         * the router outlet.
+         */
+        tabState.hasRouterOutlet = hasRouterOutlet;
+
+        /**
+         * Passing this prop to each tab button
          * lets it be aware of the state that
          * ion-tab-bar is managing for it.
          */
         child.component.props._getTabState = () => tabState;
+
+        /**
+         * If the router outlet is not defined, then the tabs are being used
+         * as a basic tab navigation without the router. In this case, the
+         * tabs will not emit the `ionTabsDidChange` and `ionTabsWillChange`
+         * events through the `checkActiveTab` method. Instead, we need to
+         * handle those events through the tab buttons.
+         */
+        if (!hasRouterOutlet) {
+          child.component.props._onClick = (
+            event: CustomEvent<{
+              href: string;
+              selected: boolean;
+              tab: string;
+            }>
+          ) => {
+            this.handleIonTabButtonClick(event);
+          };
+        }
       });
 
       this.checkActiveTab(ionRouter);
     },
+    /**
+     * This method is called upon setup and when the
+     * history changes. It checks the current route
+     * and updates the active tab accordingly.
+     *
+     * History changes only occur when the router
+     * outlet is present. Due to this, the
+     * `ionTabsDidChange` and `ionTabsWillChange`
+     * events are only emitted when the router
+     * outlet is present. A different approach must
+     * be taken for tabs without a router outlet.
+     *
+     * @param ionRouter
+     */
     checkActiveTab(ionRouter: any) {
-      const currentRoute = ionRouter.getCurrentRouteInfo();
+      const hasRouterOutlet = this.$props._hasRouterOutlet;
+      const currentRoute = ionRouter?.getCurrentRouteInfo();
       const childNodes = this.$data.tabVnodes;
       const { tabs, activeTab: prevActiveTab } = this.$data.tabState;
-      const tabState = this.$data.tabState;
       const tabKeys = Object.keys(tabs);
-      const activeTab = tabKeys.find((key) => {
+      let activeTab = tabKeys.find((key) => {
         const href = tabs[key].originalHref;
-        return currentRoute.pathname.startsWith(href);
+        return currentRoute?.pathname.startsWith(href);
       });
+
+      /**
+       * Tabs is being used as a basic tab navigation,
+       * so we need to set the first tab as active since
+       * `checkActiveTab` will not be called after setup.
+       */
+      if (!activeTab && !hasRouterOutlet) {
+        activeTab = tabKeys[0];
+      }
 
       /**
        * For each tab, check to see if the
@@ -118,17 +170,20 @@ export const IonTabBar = defineComponent({
          * If we went to tab2 then back to tab1, we should
          * land on /tabs/tab1/child instead of /tabs/tab1.
          */
-        if (activeTab !== prevActiveTab || prevHref !== currentRoute.pathname) {
+        if (
+          activeTab !== prevActiveTab ||
+          prevHref !== currentRoute?.pathname
+        ) {
           /**
            * By default the search is `undefined` in Ionic Vue,
            * but Vue Router can set the search to the empty string.
            * We check for truthy here because empty string is falsy
            * and currentRoute.search cannot ever be a boolean.
            */
-          const search = currentRoute.search ? `?${currentRoute.search}` : "";
+          const search = currentRoute?.search ? `?${currentRoute.search}` : "";
           tabs[activeTab] = {
             ...tabs[activeTab],
-            currentHref: currentRoute.pathname + search,
+            currentHref: currentRoute?.pathname + search,
           };
         }
 
@@ -137,7 +192,7 @@ export const IonTabBar = defineComponent({
          * set the previous tab back to its original href.
          */
         if (
-          currentRoute.routerAction === "pop" &&
+          currentRoute?.routerAction === "pop" &&
           activeTab !== prevActiveTab
         ) {
           tabs[prevActiveTab] = {
@@ -147,6 +202,24 @@ export const IonTabBar = defineComponent({
         }
       }
 
+      this.tabSwitch(activeTab, ionRouter);
+    },
+    handleIonTabButtonClick(
+      event: CustomEvent<{
+        href: string;
+        selected: boolean;
+        tab: string;
+      }>
+    ) {
+      const activeTab = event.detail.tab;
+
+      this.tabSwitch(activeTab);
+    },
+    tabSwitch(activeTab: string, ionRouter?: any) {
+      const hasRouterOutlet = this.$props._hasRouterOutlet;
+      const childNodes = this.$data.tabVnodes;
+      const { activeTab: prevActiveTab } = this.$data.tabState;
+      const tabState = this.$data.tabState;
       const activeChild = childNodes.find(
         (child: VNode) => isTabButton(child) && child.props?.tab === activeTab
       );
@@ -156,28 +229,31 @@ export const IonTabBar = defineComponent({
         if (activeChild) {
           tabDidChange && this.$props._tabsWillChange(activeTab);
 
-          ionRouter.handleSetCurrentTab(activeTab);
+          if (hasRouterOutlet && ionRouter !== null) {
+            ionRouter.handleSetCurrentTab(activeTab);
+          }
+
           tabBar.selectedTab = tabState.activeTab = activeTab;
 
           tabDidChange && this.$props._tabsDidChange(activeTab);
+        } else {
           /**
            * When going to a tab that does
            * not have an associated ion-tab-button
            * we need to remove the selected state from
            * the old tab.
            */
-        } else {
           tabBar.selectedTab = tabState.activeTab = "";
         }
       }
     },
   },
   mounted() {
-    const ionRouter: any = inject("navManager");
+    const ionRouter: any = inject("navManager", null);
 
     this.setupTabState(ionRouter);
 
-    ionRouter.registerHistoryChangeListener(() =>
+    ionRouter?.registerHistoryChangeListener(() =>
       this.checkActiveTab(ionRouter)
     );
   },